Stage 8 - applying the SOLID Dependency Inversion Principle part 2

+++ Homework1/src/ro/ase/cts/g1098/homework/interfaces/IRateable.java
@@ -6,7 +6,7 @@ public interface IRateable {
 	// must have method - the lead has requested it in all classes
 	public abstract double getMonthlyRate(double loanValue, double rate);
 
-	public abstract double getAnnualRate(int daysYears, double rate, int daysActive);
+	public abstract double getAnnualRate(int daysYears, double rate, double daysActive);
 	
 	public abstract void showRate(double rate, String rateType);
